+++ kwcoco/__init__.py
@@ -259,9 +259,12 @@ Some of the above docs were generated via:
 
 The logic of this init is generated via:
     mkinit ~/code/kwcoco/kwcoco/__init__.py
+
+    mkinit ~/code/kwcoco/kwcoco/__init__.py --lazy
 """
 
-__version__ = '0.2.25'
+__version__ = '0.2.26'
+
 
 __submodules__ = {
     'abstract_coco_dataset': ['AbstractCocoDataset'],
@@ -270,6 +273,11 @@ __submodules__ = {
     'category_tree': ['CategoryTree'],
     'channel_spec': ['ChannelSpec', 'FusedChannelSpec'],
     'exceptions': [],
+}
+
+__lazy_submodules__ = {
+    # TODO: always lazy submodules
+    'coco_sql_dataset': ['CocoSqlDatabase'],
 }
 
 from kwcoco import abstract_coco_dataset
@@ -285,7 +293,56 @@ from kwcoco.coco_image import (CocoImage,)
 from kwcoco.category_tree import (CategoryTree,)
 from kwcoco.channel_spec import (ChannelSpec, FusedChannelSpec,)
 
+
+def lazy_import(module_name, submodules, submod_attrs):
+    import importlib
+    import os
+    name_to_submod = {
+        func: mod for mod, funcs in submod_attrs.items()
+        for func in funcs
+    }
+
+    def __getattr__(name):
+        if name in submodules:
+            attr = importlib.import_module(
+                '{module_name}.{name}'.format(
+                    module_name=module_name, name=name)
+            )
+        elif name in name_to_submod:
+            submodname = name_to_submod[name]
+            module = importlib.import_module(
+                '{module_name}.{submodname}'.format(
+                    module_name=module_name, submodname=submodname)
+            )
+            attr = getattr(module, name)
+        else:
+            raise AttributeError(
+                'No {module_name} attribute {name}'.format(
+                    module_name=module_name, name=name))
+        globals()[name] = attr
+        return attr
+
+    if os.environ.get('EAGER_IMPORT', ''):
+        for name in name_to_submod.values():
+            __getattr__(name)
+
+        for attrs in submod_attrs.values():
+            for attr in attrs:
+                __getattr__(attr)
+    return __getattr__
+
+
+__getattr__ = lazy_import(
+    __name__,
+    submodules=[],
+    submod_attrs=__lazy_submodules__,
+)
+
+
+def __dir__():
+    return __all__
+
 __all__ = ['AbstractCocoDataset', 'CategoryTree', 'ChannelSpec', 'CocoDataset',
-           'CocoImage', 'FusedChannelSpec', 'abstract_coco_dataset',
-           'category_tree', 'channel_spec', 'coco_dataset', 'coco_image',
-           'exceptions']
+           'CocoImage', 'CocoSqlDatabase', 'FusedChannelSpec',
+           'abstract_coco_dataset', 'category_tree', 'channel_spec',
+           'coco_dataset', 'coco_image', 'coco_sql_dataset', 'exceptions']