+++ kwcoco/coco_objects1d.py
@@ -460,7 +460,7 @@ class Images(ObjectList1D):
             >>> print(ub.repr2(self.n_annots, nl=0))
             [9, 2, 0]
         """
-        return list(map(len, ub.take(self._dset.gid_to_aids, self._ids)))
+        return list(map(len, ub.take(self._dset.index.gid_to_aids, self._ids)))
 
     @property
     def aids(self):
@@ -471,7 +471,7 @@ class Images(ObjectList1D):
             >>> print(ub.repr2(list(map(list, self.aids)), nl=0))
             [[1, 2, 3, 4, 5, 6, 7, 8, 9], [10, 11], []]
         """
-        return list(ub.take(self._dset.gid_to_aids, self._ids))
+        return list(ub.take(self._dset.index.gid_to_aids, self._ids))
 
     @property
     def annots(self):