+++ CHANGELOG.md
@@ -5,7 +5,23 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.25 - Unreleased
+## Version 0.2.26 - Unreleased
+
+### Added
+
+* Method to get the valid region for a CocoImage in image or video space
+
+
+### Changed
+
+* CocoSqlDatabase is now lazy exposed at the top level in Python 3.7+
+* Better SQL support in CocoImage
+* Add maxlen in `path_sanitize`
+* Warp properties in CocoImage
+* Stub out CocoAsset
+
+
+## Version 0.2.25 - Released 2022-03-16
 
 ### Added
 * Can now load do delayed loading of data in auxiliary space. (throws an error